+++ src/config.h
@@ -25,6 +25,7 @@ enum class ConfigSource {
 enum class ConfigKeyBool {
   cpp_demangle,
   lazy_symbolication,
+  probe_inline,
 };
 
 enum class ConfigKeyInt {
@@ -71,6 +72,7 @@ const std::map<std::string, ConfigKey> CONFIG_KEY_MAP = {
   { "max_strlen", ConfigKeyInt::max_strlen },
   { "max_type_res_iterations", ConfigKeyInt::max_type_res_iterations },
   { "perf_rb_pages", ConfigKeyInt::perf_rb_pages },
+  { "probe_inline", ConfigKeyBool::probe_inline },
   { "stack_mode", ConfigKeyStackMode::default_ },
   { "str_trunc_trailer", ConfigKeyString::str_trunc_trailer },
 };